exercise 4.23

+++ part4/bloglist-backend/controllers/blogs.js
@@ -39,6 +39,16 @@ blogsRouter.post('/', async (request, response) => {
 
 // Borrar un blog
 blogsRouter.delete('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+
+  if (!request.user || blog.user.toString() !== request.user.id.toString()) {
+    return response.status(401).json({ error: 'unauthorized: not owner' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
